requirements.txt
crewai
google-auth
google-auth-oauthlib
google-api-python-client
google-auth-httplib2